requirements.txt
biopython >= 1.79
cutadapt == 3.7
pandas >= 1.1.5
scipy >= 1.5.4
python-Levenshtein >= 0.12.2
matplotlib >= 3.3.4
medaka >= 1.5
